Add a typeResults capability to simulate a program running
We add a text file to complement the .cs file, for example hello-world.cs has hello-world.txt
This is typed in the console area
Lines starting with a *, are user responses
We add the ability to pick which file to run by specifying in the URL using ?file=hello-world.cs
We add the ability to change typing speed using URL parameter speedUpFactor. e.g. http://127.0.0.1:5500/?file=rpg-inventory.cs&speedUpFactor=100. This defaults to 1.